requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
celery==5.6.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
django==6.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
exceptiongroup==1.3.1
factory-boy==3.3.3
faker==38.2.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
packaging==25.0
pluggy==1.6.0
prompt-toolkit==3.0.52
psycopg2==2.9.11
pygments==2.19.2
pytest==9.0.1
pytest-django==4.11.1
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
pyyaml==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
